PRELIM STEP before risky change ...
Change all appropriate 'opponent' vars with 'XYZXYZXYZ'
for easy identification, so that, in the event of a problem with
the follow up change in the next step (FOLLOW UP), it will be easier to
troubleshoot the problem introduced with the next change.

FOLLOW UP: The next change will be to replace all 'XYZXYZXYZ' with 'avdersaryID'
as a counterpart nomenclature to 'registrationID' for the player.